backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic[email]==2.10.0
pydantic-settings==2.6.0

# Supabase
supabase==2.11.0

# OpenAI
openai==1.58.0

# Image processing
Pillow==11.0.0
opencv-python-headless==4.10.0.84
mediapipe==0.10.32

# Payment processing
stripe==11.4.0

# Replicate (FLUX Kontext Pro)
replicate==1.0.4

# Google Gemini
google-genai==1.0.0

# Anthropic Claude
anthropic==0.40.0

# Human body parsing (fashn-ai/fashn-human-parser)
torch>=2.2.0
transformers>=4.39.0

# Local CPU diffusion inpainting
diffusers>=0.31.0
accelerate>=0.33.0
safetensors>=0.4.0
huggingface-hub>=0.25.0
scipy>=1.12.0

# HTTP requests
httpx==0.28.0
svix==1.66.0

# Date/time utilities
python-dateutil==2.9.0

# Rate limiting
slowapi==0.1.9

# Tests (async + pytest)
pytest>=8.0.0
pytest-asyncio>=0.24.0
